+++ packages/git-packet-streams/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.2.0] - 2019-02-06
+
+### Changed
+
+- All instances of `Buffer`, except for in the PacketReader is replaced with its
+  super-class `Uint8Array`.
+
 ## [2.1.1] - 2018-12-30
 
 ### Fixed
@@ -102,7 +109,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
-[Unreleased]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.1.1...HEAD
+[Unreleased]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.2.0...HEAD
+[2.2.0]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.1.1...git-packet-streams-v2.2.0
 [2.1.1]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.1.0...git-packet-streams-v2.1.1
 [2.1.0]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.0.2...git-packet-streams-v2.1.0
 [2.0.2]: https://github.com/revam/node-git-monorepo/compare/git-packet-streams-v2.0.1...git-packet-streams-v2.0.2